+++ keys_test.go
@@ -63,7 +63,7 @@ func TestParseDelegationsFromResponse(t *testing.T) {
 			name:        "fail - no keys",
 			out:         "invalid output",
 			expError:    true,
-			errContains: "error unmarshalling accounts",
+			errContains: "error unmarshalling delegations",
 		},
 	}