+++ proposal.go
@@ -2,8 +2,18 @@ package main
 
 import (
 	"fmt"
-	"regexp"
+	abcitypes "github.com/cometbft/cometbft/abci/types"
+	"github.com/evmos/evmos/v14/app"
+	"github.com/evmos/evmos/v14/encoding"
 	"strconv"
+	"strings"
+)
+
+var (
+	// cdc is the codec to be used for the client
+	cdc = encodingConfig.Codec
+	// encodingConfig specifies the encoding configuration to be used for the client
+	encodingConfig = encoding.MakeConfig(app.ModuleBasics)
 )
 
 // submitUpgradeProposal submits a software upgrade proposal with the given target version and upgrade height.
@@ -14,23 +24,38 @@ func submitUpgradeProposal(targetVersion string, upgradeHeight int) (int, error)
 		return 0, err
 	}
 
-	return getProposalID(out)
-}
+	// Clean gas estimate output and only leave json output
+	out = strings.TrimSpace(out)
+	lines := strings.Split(out, "\n")
+	out = lines[len(lines)-1] // last line is json output
 
-// getProposalID parses the proposal ID from the given output from submitting an upgrade proposal.
-func getProposalID(out string) (int, error) {
-	// Define the regular expression pattern
-	pattern := `- key:\s*proposal_id\s*\n\s*value:\s*"([^"]+)"`
-
-	// Compile the regular expression
-	re := regexp.MustCompile(pattern)
-
-	match := re.FindStringSubmatch(out)
-	if len(match) != 2 {
-		return 0, fmt.Errorf("proposal ID not found in output")
+	events, err := getTxEvents(out)
+	if err != nil {
+		panic(err)
 	}
 
-	return strconv.Atoi(match[1])
+	return getProposalID(events)
+}
+
+// getProposalID looks for the proposal submission event in the given transaction events
+// and returns the proposal id, if found.
+func getProposalID(events []abcitypes.Event) (int, error) {
+	for _, event := range events {
+		if event.Type != "submit_proposal" {
+			continue
+		}
+		for _, attribute := range event.Attributes {
+			if attribute.Key == "proposal_id" {
+				proposalID, err := strconv.Atoi(attribute.Value)
+				if err != nil {
+					return 0, fmt.Errorf("error parsing proposal id: %w", err)
+				}
+				return proposalID, nil
+			}
+		}
+	}
+
+	return 0, fmt.Errorf("proposal submission event not found")
 }
 
 // buildUpgradeProposalCommand builds the command to submit a software upgrade proposal.
@@ -41,6 +66,7 @@ func buildUpgradeProposalCommand(targetVersion string, upgradeHeight int) []stri
 		"--description", fmt.Sprintf("'Upgrade to %s'", targetVersion),
 		"--upgrade-height", fmt.Sprintf("%d", upgradeHeight),
 		"--deposit", "100000000000000000000aevmos",
+		"--output", "json",
 		"--no-validate",
 	}
 }