+++ CHANGELOG.md
@@ -35,11 +35,17 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
+## Unreleased
+
+### Improvements
+
+- [#3](https://github.com/MalteHerrmann/upgrade-local-node-go/pull/3) Use broadcast mode `sync` instead of `block`
+
 ## [v0.2.0](https://github.com/MalteHerrmann/upgrade-local-node-go/releases/tag/v0.2.0) - 2023-08-09
 
 ### Improvements
 
-- [#1](https://github.com/MalteHerrmann/upgrade-local-node-go/pull/1) adaptively gets keys and current proposal ID from the local node
+- [#1](https://github.com/MalteHerrmann/upgrade-local-node-go/pull/1) Adaptively gets keys and current proposal ID from the local node
 
 ## [v0.1.0](https://github.com/MalteHerrmann/upgrade-local-node-go/releases/tag/v0.1.0) - 2023-08-01