+++ CHANGELOG.md
@@ -40,12 +40,14 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Improvements
 
 - [#2](https://github.com/MalteHerrmann/upgrade-local-node-go/pull/2) only vote if account has delegations
+- [#3](https://github.com/MalteHerrmann/upgrade-local-node-go/pull/3) Use broadcast mode `sync` instead of `block`
+- [#4](https://github.com/MalteHerrmann/upgrade-local-node-go/pull/4) Add GH actions and Makefile for testing
 
 ## [v0.2.0](https://github.com/MalteHerrmann/upgrade-local-node-go/releases/tag/v0.2.0) - 2023-08-09
 
 ### Improvements
 
-- [#1](https://github.com/MalteHerrmann/upgrade-local-node-go/pull/1) adaptively gets keys and current proposal ID from the local node
+- [#1](https://github.com/MalteHerrmann/upgrade-local-node-go/pull/1) Adaptively gets keys and current proposal ID from the local node
 
 ## [v0.1.0](https://github.com/MalteHerrmann/upgrade-local-node-go/releases/tag/v0.1.0) - 2023-08-01